Refactored DFS
Removed the horrible recursive function where we passed in lots of
state. Replaced it with a recursive function implemented on Graph
itself. This means we only need to pass along to the next stage of the
recursion the variables that are describing the recursion, because it
already has access to all of the variables about the Graph.